Selenium grid

+++ src/test/java/testBase/Baseclass.java
@@ -2,16 +2,20 @@ package testBase;
 
 import java.io.FileInputStream;
 import java.io.IOException;
+import java.net.URL;
 import java.time.Duration;
 import java.util.Properties;
 
 import org.apache.commons.lang3.RandomStringUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.openqa.selenium.Platform;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.edge.EdgeDriver;
 import org.openqa.selenium.firefox.FirefoxDriver;
+import org.openqa.selenium.remote.DesiredCapabilities;
+import org.openqa.selenium.remote.RemoteWebDriver;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Parameters;
@@ -27,11 +31,26 @@ public class Baseclass {
 		FileInputStream file=new FileInputStream("./src//test//resources//config.properties");
 		p=new Properties();
 		p.load(file);
-		switch (br.toLowerCase()) {
-		case "firefox": driver=new FirefoxDriver();break;
-		case "edge": driver=new EdgeDriver();break;
-		case "chrome": driver=new ChromeDriver();break;
+		if(p.getProperty("execution_env").equalsIgnoreCase("remote")) {
+		DesiredCapabilities cp=new DesiredCapabilities();
+			//ChromeOptions CO = new ChromeOptions();
+			//CO.setPlatformName(os);
+			//CO.setBrowserVersion(br);
+			cp.setPlatform(Platform.WIN11);
+			cp.setBrowserName("chrome");
+			driver=new RemoteWebDriver(new URL ("http://localhost:4444/"),cp);
+	}
+		
+	
+		if(p.getProperty("execution_env").equalsIgnoreCase("local")) {
+			switch (br.toLowerCase()) {
+			case "firefox": driver=new FirefoxDriver();break;
+			case "edge": driver=new EdgeDriver();break;
+			case "chrome": driver=new ChromeDriver();break;
+			default: System.out.println("Invalid browser name..");return;
+			}
 		}
+		
 		driver.manage().deleteAllCookies();
 		driver.manage().window().maximize();
 		driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(15));
